♻️🏢 Split new FF route
