+++ apps/admin/src/routes/_layout.foster-families.new/loader.server.ts
@@ -1,0 +1,18 @@
+import { db } from "#core/db.server";
+import { ok } from "#core/response.server.js";
+import { assertCurrentUserHasGroups } from "#current-user/groups.server";
+import { UserGroup } from "@prisma/client";
+import type { LoaderFunctionArgs } from "@remix-run/node";
+
+export async function loader({ request }: LoaderFunctionArgs) {
+  const currentUser = await db.currentUser.get(request, {
+    select: { groups: true },
+  });
+
+  assertCurrentUserHasGroups(currentUser, [
+    UserGroup.ADMIN,
+    UserGroup.ANIMAL_MANAGER,
+  ]);
+
+  return ok();
+}
